lvol: allow to configure cluster size of lvol store
New optional parameter -s or --cluster_sz in construct_lvol_store() RPC,
as well as in vbdev_lvol and lvol lib API.

This parameter allows to configure cluster size of blobstore that lvol store
is build upon.

When this parameter is not specified, default of 1GiB is used.

spdk_lvs_opts struct was created to facilitate any future options when
creating lvol store.

Any occurrence of cluster size in blobstore was change from uint32_t
to uint64_t, in order to enable cluster sizes above 2GiB.

+++ include/spdk/blob.h
@@ -140,7 +140,7 @@ struct spdk_bs_dev {
 };
 
 struct spdk_bs_opts {
-	uint32_t cluster_sz; /* In bytes. Must be multiple of page size. */
+	uint64_t cluster_sz; /* In bytes. Must be multiple of page size. */
 	uint32_t num_md_pages; /* Count of the number of pages reserved for metadata */
 	uint32_t max_md_ops; /* Maximum simultaneous metadata operations */
 	uint32_t max_channel_ops; /* Maximum simultaneous operations per channel */
